+++ src/to-do-lists/to-do-lists.controller.ts
@@ -1,29 +1,52 @@
-import { Body, Controller, Delete, Get, Param, Post } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Post,
+  Req,
+} from '@nestjs/common';
 import { ToDoListsService } from './to-do-lists.service';
 import { ToDoList as ToDoListModel } from '@prisma/client';
+import { Request } from 'express';
 
 @Controller('lists')
 export class ToDoListsController {
   constructor(private readonly toDoListsService: ToDoListsService) {}
   @Get()
-  async getLists(): Promise<ToDoListModel[]> {
-    return this.toDoListsService.toDoLists({});
+  async getLists(@Req() req: Request): Promise<ToDoListModel[]> {
+    const { sub: userId } = req.user;
+
+    return this.toDoListsService.toDoLists({ where: { authorId: userId } });
   }
 
   @Get(':id')
-  async getListById(@Param('id') id: string): Promise<ToDoListModel> {
-    return this.toDoListsService.toDoList({ id: id });
+  async getListById(
+    @Req() req: Request,
+    @Param('id') id: string,
+  ): Promise<ToDoListModel> {
+    const { sub: userId } = req.user;
+    return this.toDoListsService.toDoList(userId, { id: id });
   }
 
   @Post('new')
   async create(
+    @Req() req: Request,
     @Body() createToDoListDto: ToDoListModel,
   ): Promise<ToDoListModel> {
-    return this.toDoListsService.createToDoList(createToDoListDto);
+    const { sub: userId } = req.user;
+
+    return this.toDoListsService.createToDoList(userId, createToDoListDto);
   }
 
   @Delete(':id')
-  async deletePost(@Param('id') id: string): Promise<ToDoListModel> {
-    return this.toDoListsService.deleteToDoList({ id });
+  async deletePost(
+    @Req() req: Request,
+    @Param('id') id: string,
+  ): Promise<ToDoListModel> {
+    const { sub: userId } = req.user;
+
+    return this.toDoListsService.deleteToDoList(userId, { id });
   }
 }
